+++ ecaudit/src/main/java/com/ericsson/bss/cassandra/ecaudit/entry/CqlLiteralVersionAdapter.java
@@ -23,6 +23,7 @@ import java.util.TimeZone;
 
 import org.apache.cassandra.cql3.ColumnSpecification;
 import org.apache.cassandra.db.marshal.AsciiType;
+import org.apache.cassandra.db.marshal.BytesType;
 import org.apache.cassandra.db.marshal.TimestampType;
 import org.apache.cassandra.db.marshal.UTF8Type;
 import org.apache.cassandra.serializers.TimestampSerializer;
@@ -54,6 +55,10 @@ public final class CqlLiteralVersionAdapter
         {
             return DATE_FORMAT.format(TimestampSerializer.instance.deserialize(serializedValue));
         }
+        if (isBlobType(column))
+        {
+            return "0x" + column.type.getString(serializedValue);
+        }
 
         return column.type.getString(serializedValue);
     }
@@ -71,6 +76,11 @@ public final class CqlLiteralVersionAdapter
     private static boolean isTimestampType(ColumnSpecification column)
     {
         return column.type instanceof TimestampType;
+    }
+
+    private static boolean isBlobType(ColumnSpecification column)
+    {
+        return column.type instanceof BytesType;
     }
 
     private static DateFormat createDateFormat()